feat: arquivos flyway migration adicionado para tabelas de admin e status pagamento

+++ src/main/resources/db/migration/V2__create_table_pagamento.sql
@@ -1,6 +1,6 @@
 CREATE TABLE pagamento (
-    id_pagamento BIGINT AUTO_INCREMENT PRIMARY KEY,
-    id_pedido BIGINT NOT NULL,
+    id_pagamento INT AUTO_INCREMENT PRIMARY KEY,
+    id_pedido INT NOT NULL,
     valor_total DECIMAL(10,2) NOT NULL,
     metodo_pagamento VARCHAR(100) NOT NULL,
     id_status_pagamento BIGINT NOT NULL,
